+++ CodexLib/Patches/Patch_SuppressError.cs
@@ -6,6 +6,11 @@ namespace CodexLib
     public class Patch_SuppressError
     {
         [HarmonyPatch(typeof(UnitPartKineticist), nameof(UnitPartKineticist.MaxBurn), MethodType.Getter)]
-        public static Exception Finalizer() => null;
+        [HarmonyFinalizer]
+        public static Exception Finalizer1() => null;
+
+        [HarmonyPatch(typeof(UnitPartKineticist), nameof(UnitPartKineticist.GatherPowerAbility), MethodType.Getter)]
+        [HarmonyFinalizer]
+        public static Exception Finalizer2() => null;
     }
 }